For Tables.columns fallback, attempt to preserve schema on empty inputs (#206)
* For Tables.columns fallback, attempt to preserve schema on empty inputs

Fixes #204. For empty row interator inputs, we were just returning an
empty `NamedTuple`. This has the disadvantage of not preserving an
input's schema in the case like `NamedTuple{(:a, :b), Tuple{Int64,
Float64}}[]`. Sometimes the input may be empty, but have a queryable
schema, so we should try to preserve that. For `Tuple` rows, we generate
column names like `Column$i`. I think this is fine because we're in the
fallback `buildcolumns` code where we just want to return a standard
"table" object, so returning a NamedTuple instead of `Tuple` of vectors
seems more appropriate; i.e. it's Tables.jl job here to "build the
columns", so we have latitude and control over what we return.