Introduce a specialization threshold and try to avoid generated code
Vastly improves #209. I actually thought this would trip us up sooner,
but kind of forgot about it. We had some generous use of `@generated`
code in a number of places, most aggregiously in
`Tables.allocatecolumns` and `Tables.eachcolumn`. It simplifies the code
to either just use a slow fallback, or skip the generated code entirely
by using macro expansion up to a certain limit. I tried to test out a
number of cases, comparing time to compile vs. resulting compiled code
performance, but it can be tricky to cover a wide variety of workflows.
I might try to generate a wider variety to compare benchmarks pre/post
this PR.