Add dictcolumntable and dictrowtable functions
Fixes #161 and #207. This defines two table types, `DictColumnTable` and
`DictRowTable`, along wtih corresponding constructors, `dictcolumntable`
and `dictrowtable`. For #161, both table types hold their own schema,
which provides a way to specify the "ordering" of the tables, even
though the underlying data is stored in `Dict`s. For #207, both these
table types employ a "column unioning" approach for schema-less input
row iterators. This is similar to the `cols=:union` behavior you get
with DataFrames `push!`, and at least provides a way for users to get
this behavior when so desired. I like this approach because to even
implement the functionality, you basically need to treat rows/columns as
`Dict`s, to check if keys are present or not, and they're very flexible
in terms of adding additional key-value pairs (much easier than trying
to "update" `NamedTuple`s). So this is a bit of a compromise: we keep
the current great performance of `Tables.columns` fallback via
`Tables.buildcolumns` for schema-less row iterator inputs, but provide a
pretty simple "intermediate" table type that will union columns if
needed, via `dictcolumntable` and `dictrowtable`.