Don't lie when calling Tables.Row/Tables.Columns
Fixes #254. The problem with returning the input itself for any
`::AbstractColumns` is that it makes it hard to use `Tables.Columns` as
a dispatchable type when desired. If someone passes an object that's
already `::AbstractColumns`, it's also not the end of the world to wrap
it in one more layer where `getcolumn` calls are probably going to get
inlined away anyway.