Fix inconsistency between row and column behavior of MatrixTable (#262)
Fixes #261. Because `MatrixTable` subtypes `AbstractColumns`, it
inherits certain behavior like `length`; but when it needs to fulfill
the "rows" interface, it needs `length` to be defined in terms of the #
of rows instead of the # of columns. We fix this inconsistency here by
defining another type that mirrors `MatrixTable` but is used in the rows
case.