Fix columntable materialization on stored schema
Fixes https://github.com/JuliaData/Tables.jl/issues/357.

The issue here is for stored schema, the type of the schema is
`Schema{nothing, nothing}` which usually indicates tables with
many columns. Some tables implementations, however, like ARFFFiles.jl,
may choose to explicitly store _all_ schemas, even for very narrow tables.
We already have a generated branch which checks for a specialization threshold
for the known-schema case, so the fix here is fairly straightforward in just
actually checking if the stored schema # of columns is actually too many or
not.

In the end, users should be aware that `Tables.columntable` isn't a perfect,
100% kind of table implementation that is always expected to work. It was originally
meant as just a test implementation that then turned out to be fairly convenient
for REPL use. Users should note that generating a named tuple of columns from stored
schema doesn't have a way to be particularly efficient, since it necessarily has to
generate the NamedTuple type at runtime.